+++ lib/presentation/viewmodels/home/tabs/home_tab_viewmodel.dart
@@ -3,8 +3,8 @@ import 'package:roadmap/domain/entities/goal_item.dart';
 import 'package:roadmap/domain/usecases/goal_item_usecase.dart';
 import 'package:roadmap/presentation/routes/navigation_service.dart';
 
-class HomeTabViewmodel extends StateNotifier<AsyncValue<List<GoalItem>>> {
-  HomeTabViewmodel(this._navigationService, this._usecase)
+class HomeTabViewModel extends StateNotifier<AsyncValue<List<GoalItem>>> {
+  HomeTabViewModel(this._navigationService, this._usecase)
       : super(const AsyncValue.loading()) {
     retrieveItems();
   }
@@ -31,15 +31,13 @@ class HomeTabViewmodel extends StateNotifier<AsyncValue<List<GoalItem>>> {
     }
     try {
       final items = await _usecase.retrieveItems();
-      if (mounted) {
-        state = AsyncValue.data(
-          _filter == null
-              ? items
-              : items.where((item) => item.isCompleted == _filter).toList(),
-        );
-      }
-    } on Exception {
-      // NOP
+      state = AsyncValue.data(
+        _filter == null
+            ? items
+            : items.where((item) => item.isCompleted == _filter).toList(),
+      );
+    } on Exception catch (e) {
+      state = AsyncValue.error(e.toString(), StackTrace.empty);
     }
   }
 
@@ -60,21 +58,22 @@ class HomeTabViewmodel extends StateNotifier<AsyncValue<List<GoalItem>>> {
         createdAt: DateTime.now(),
       );
       final itemId = await _usecase.createItem(item);
-      state.whenData(
-        (items) => state = AsyncValue.data(
-          items..add(item.copyWith(id: itemId)),
-        ),
+      state.maybeWhen(
+        data: (items) {
+          state = AsyncValue.data(items..add(item.copyWith(id: itemId)));
+        },
+        orElse: () {},
       );
-    } on Exception {
-      // NOP
+    } on Exception catch (e) {
+      state = AsyncValue.error(e.toString(), StackTrace.empty);
     }
   }
 
   Future<void> updateItem({required GoalItem updatedItem}) async {
     try {
       await _usecase.updateItem(updatedItem);
-      state.whenData(
-        (items) {
+      state.maybeWhen(
+        data: (items) {
           state = AsyncValue.data(
             [
               for (final item in items)
@@ -82,22 +81,26 @@ class HomeTabViewmodel extends StateNotifier<AsyncValue<List<GoalItem>>> {
             ],
           );
         },
+        orElse: () {},
       );
-    } on Exception {
-      // NOP
+    } on Exception catch (e) {
+      state = AsyncValue.error(e.toString(), StackTrace.empty);
     }
   }
 
   Future<void> deleteItem({required String itemId}) async {
     try {
       await _usecase.deleteItem(itemId);
-      state.whenData(
-        (items) => state = AsyncValue.data(
-          items..removeWhere((item) => item.id == itemId),
-        ),
+      state.maybeWhen(
+        data: (items) {
+          state = AsyncValue.data(
+            items..removeWhere((item) => item.id == itemId),
+          );
+        },
+        orElse: () {},
       );
-    } on Exception {
-      // NOP
+    } on Exception catch (e) {
+      state = AsyncValue.error(e.toString(), StackTrace.empty);
     }
   }
 }
