+++ lib/presentation/views/home/home_screen.dart
@@ -5,7 +5,6 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:roadmap/core/di/providers.dart';
 import 'package:roadmap/presentation/views/home/tabs/home_tab.dart';
 import 'package:roadmap/presentation/views/home/tabs/stats_tab.dart';
-import 'package:roadmap/presentation/views/splash/splash_screen.dart';
 import 'package:roadmap/presentation/views/welcome/welcome_screen.dart';
 
 class HomeScreen extends HookConsumerWidget {
@@ -18,23 +17,10 @@ class HomeScreen extends HookConsumerWidget {
     final borderRadiusAnimationController = useAnimationController(
       duration: const Duration(milliseconds: 500),
     )..forward();
-
-    Widget tabContent() {
-      switch (tabIndex.value) {
-        case 0:
-          return const HomeTab();
-        case 1:
-          return const StatsTab();
-        default:
-          return Container();
-      }
-    }
-
-    final iconList = <IconData>[
+    final iconList = [
       Icons.home,
       Icons.bar_chart,
     ];
-
     final authStateAsyncValue = ref.watch(authStateProvider);
 
     return authStateAsyncValue.when(
@@ -42,56 +28,86 @@ class HomeScreen extends HookConsumerWidget {
         if (user != null) {
           return Scaffold(
             extendBody: true,
-            body: tabContent(),
-            floatingActionButton: tabIndex.value != 1
-                ? FloatingActionButton(
-                    onPressed: viewModel.navigateToAdd,
-                    child: const Icon(Icons.add),
-                  )
-                : null,
+            body: _tabContent(index: tabIndex.value),
+            floatingActionButton: _buildFloatingActionButton(
+              tabIndex.value,
+              viewModel.navigateToAdd,
+            ),
             floatingActionButtonLocation:
                 FloatingActionButtonLocation.centerDocked,
-            bottomNavigationBar: AnimatedBottomNavigationBar.builder(
-              itemCount: iconList.length,
-              tabBuilder: (int index, bool isActive) {
-                final color = isActive ? Colors.white : Colors.grey;
-                return Column(
-                  mainAxisSize: MainAxisSize.min,
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Icon(
-                      iconList[index],
-                      size: 24,
-                      color: color,
-                    ),
-                    const SizedBox(height: 4),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 8),
-                      child: Text(
-                        index == 0 ? 'Home' : 'Stats',
-                        style: TextStyle(color: color),
-                      ),
-                    ),
-                  ],
-                );
-              },
-              backgroundColor: Colors.grey[900],
-              activeIndex: tabIndex.value,
-              splashColor: Colors.white,
-              notchAndCornersAnimation: borderRadiusAnimationController,
-              splashSpeedInMilliseconds: 300,
-              notchSmoothness: NotchSmoothness.defaultEdge,
-              gapLocation: GapLocation.center,
-              leftCornerRadius: 32,
-              rightCornerRadius: 32,
-              onTap: (index) => tabIndex.value = index,
+            bottomNavigationBar: _buildBottomNavigationBar(
+              tabIndex,
+              borderRadiusAnimationController,
+              iconList,
             ),
           );
         }
         return const WelcomeScreen();
       },
-      loading: () => const SplashScreen(),
+      loading: () => const Center(child: CircularProgressIndicator()),
       error: (error, _) => Text(error.toString()),
+    );
+  }
+
+  Widget _tabContent({required int index}) {
+    switch (index) {
+      case 0:
+        return const HomeTab();
+      case 1:
+        return const StatsTab();
+      default:
+        return Container();
+    }
+  }
+
+  FloatingActionButton? _buildFloatingActionButton(
+    int tabIndex,
+    VoidCallback onPressed,
+  ) {
+    if (tabIndex != 1) {
+      return FloatingActionButton(
+        onPressed: onPressed,
+        child: const Icon(Icons.add),
+      );
+    }
+    return null;
+  }
+
+  AnimatedBottomNavigationBar _buildBottomNavigationBar(
+    ValueNotifier<int> tabIndex,
+    AnimationController borderRadiusAnimationController,
+    List<IconData> iconList,
+  ) {
+    return AnimatedBottomNavigationBar.builder(
+      itemCount: iconList.length,
+      tabBuilder: (int index, bool isActive) {
+        final color = isActive ? Colors.white : Colors.grey;
+        return Column(
+          mainAxisSize: MainAxisSize.min,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(iconList[index], size: 24, color: color),
+            const SizedBox(height: 4),
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 8),
+              child: Text(
+                index == 0 ? 'Home' : 'Stats',
+                style: TextStyle(color: color),
+              ),
+            ),
+          ],
+        );
+      },
+      backgroundColor: Colors.grey[900],
+      activeIndex: tabIndex.value,
+      splashColor: Colors.white,
+      notchAndCornersAnimation: borderRadiusAnimationController,
+      splashSpeedInMilliseconds: 300,
+      notchSmoothness: NotchSmoothness.defaultEdge,
+      gapLocation: GapLocation.center,
+      leftCornerRadius: 32,
+      rightCornerRadius: 32,
+      onTap: (index) => tabIndex.value = index,
     );
   }
 }
