+++ lib/presentation/widgets/tile/home_list_tile.dart
@@ -1,0 +1,77 @@
+import 'package:flutter/material.dart';
+
+class HomeListTile extends StatelessWidget {
+  const HomeListTile({
+    super.key,
+    required this.title,
+    required this.deadline,
+    required this.progress,
+    required this.imagePath,
+    required this.onTap,
+    required this.isCompleted,
+  });
+  final String title;
+  final String deadline;
+  final int progress;
+  final String imagePath;
+  final VoidCallback onTap;
+  final bool isCompleted;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        margin: const EdgeInsets.all(10),
+        padding: const EdgeInsets.all(10),
+        decoration: BoxDecoration(
+          image: DecorationImage(
+            image: AssetImage(imagePath),
+            fit: BoxFit.cover,
+          ),
+          borderRadius: BorderRadius.circular(10),
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Text(
+              title,
+              style: const TextStyle(
+                color: Colors.white,
+                fontSize: 20,
+              ),
+            ),
+            const SizedBox(height: 10),
+            Text(
+              deadline,
+              style: const TextStyle(
+                color: Colors.white70,
+                fontSize: 16,
+              ),
+            ),
+            const SizedBox(height: 10),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Expanded(
+                  child: LinearProgressIndicator(
+                    value: progress / 100,
+                    backgroundColor: Colors.grey,
+                    valueColor: const AlwaysStoppedAnimation<Color>(
+                      Colors.yellow,
+                    ),
+                  ),
+                ),
+                const SizedBox(width: 10),
+                Text(
+                  '$progress%',
+                  style: const TextStyle(color: Colors.white, fontSize: 16),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
